Redesign to allow multiple games, spellingbee, wordle
