+++ src/games/mod.rs
@@ -1,0 +1,2 @@
+pub mod bee;
+pub mod word;
